Fix manager_week/manager_year_week column mismatch causing transformation failures
- Add bidirectional alias in add_composite_keys(): if either column exists,
  create the other to ensure consistency between matchup and player data
- Update player_to_matchup_v2.py to handle either column name when validating
- Stop dropping manager_week after processing to preserve it for downstream scripts
- This fixes the root cause of 16/18 transformation failures in cloud imports

+++ analytics_app/md/tab_data_access/draft/draft_data.py
@@ -20,45 +20,37 @@ DRAFT_COLUMNS = [
     "nfl_team",
     "manager",
     "year",
-
     # Draft position & cost
     "cost",
     "pick",
     "round",
     "cost_bucket",
-
     # Keeper status & economics
     "is_keeper_status",
     "is_keeper_cost",
     "kept_next_year",
-    "keeper_price",         # Next year's keeper price (from context rules)
-    "keeper_year",          # How many consecutive years kept
-
+    "keeper_price",  # Next year's keeper price (from context rules)
+    "keeper_year",  # How many consecutive years kept
     # Performance metrics
     "total_fantasy_points",
     "season_ppg",
     "games_played",
-    "games_missed",         # Games missed (injury tracking)
+    "games_missed",  # Games missed (injury tracking)
     "weeks_rostered",
     "weeks_started",
-
     # Dual SPAR metrics (player vs manager)
-    "player_spar",      # Total SPAR (opportunity cost - what player produced)
-    "manager_spar",     # Managed SPAR (actual value - SPAR while rostered)
-    "spar",            # Legacy SPAR column
-
+    "player_spar",  # Total SPAR (opportunity cost - what player produced)
+    "manager_spar",  # Managed SPAR (actual value - SPAR while rostered)
+    "spar",  # Legacy SPAR column
     # Dual PPG metrics
-    "player_ppg",       # Player's total PPG
-    "manager_ppg",      # PPG while on manager's roster
-
+    "player_ppg",  # Player's total PPG
+    "manager_ppg",  # PPG while on manager's roster
     # Dual PGVOR metrics
-    "player_pgvor",     # Total PGVOR
-    "manager_pgvor",    # Managed PGVOR
-    "pgvor",           # Legacy PGVOR column
-
+    "player_pgvor",  # Total PGVOR
+    "manager_pgvor",  # Managed PGVOR
+    "pgvor",  # Legacy PGVOR column
     # Replacement level
     "replacement_ppg",
-
     # Value metrics
     "draft_roi",
     "spar_per_dollar",
@@ -66,68 +58,59 @@ DRAFT_COLUMNS = [
     "spar_per_round",
     "points_per_dollar",
     "points_per_pick",
-
     # Rankings
     "season_overall_rank",
     "season_position_rank",
     "price_rank_within_position",
     "pick_rank_within_position",
-
     # Efficiency metrics
     "value_over_replacement",
     "draft_position_delta",
     "pick_savings",
     "cost_savings",
-
     # === NEW: Draft Grades & Value Tiers ===
-    "spar_percentile",      # SPAR percentile within position-year (0-100)
-    "draft_grade",          # A/B/C/D/F letter grade
-    "value_tier",           # Steal/Good Value/Fair/Overpay/Bust
-
+    "spar_percentile",  # SPAR percentile within position-year (0-100)
+    "draft_grade",  # A/B/C/D/F letter grade
+    "value_tier",  # Steal/Good Value/Fair/Overpay/Bust
     # === NEW: Draft Flags & Tiers ===
-    "round_percentile",     # Percentile of round within year (0-100)
-    "draft_tier",           # Early (1-5) / Mid (6-10) / Late (11-16)
-    "is_breakout",          # Late-round pick with top finish
-    "is_bust",              # Early-round pick with bottom finish
-    "is_injured",           # Missed 50%+ of games
-    "is_injury_bust",       # Bust due to injury
+    "round_percentile",  # Percentile of round within year (0-100)
+    "draft_tier",  # Early (1-5) / Mid (6-10) / Late (11-16)
+    "is_breakout",  # Late-round pick with top finish
+    "is_bust",  # Early-round pick with bottom finish
+    "is_injured",  # Missed 50%+ of games
+    "is_injury_bust",  # Bust due to injury
     "is_performance_bust",  # Bust due to poor performance (not injury)
-
     # === NEW: Manager Draft Grades ===
-    "pick_quality_score",           # How good was this pick vs peers (0-100)
-    "pick_weight",                  # Weight for this pick (early picks weighted more)
-    "manager_draft_score",          # Manager's weighted avg score for year
-    "manager_draft_percentile",     # Grading percentile (all-time)
-    "manager_draft_percentile_year",    # Percentile within that year
-    "manager_draft_percentile_alltime", # Percentile across ALL years
-    "manager_draft_grade",          # A/B/C/D/F manager grade
-    "manager_total_spar",           # Sum of SPAR for all picks
-    "manager_avg_spar",             # Average SPAR per pick
-    "manager_hit_rate",             # % of picks with positive SPAR
-    "manager_picks_count",          # Total picks made
-
+    "pick_quality_score",  # How good was this pick vs peers (0-100)
+    "pick_weight",  # Weight for this pick (early picks weighted more)
+    "manager_draft_score",  # Manager's weighted avg score for year
+    "manager_draft_percentile",  # Grading percentile (all-time)
+    "manager_draft_percentile_year",  # Percentile within that year
+    "manager_draft_percentile_alltime",  # Percentile across ALL years
+    "manager_draft_grade",  # A/B/C/D/F manager grade
+    "manager_total_spar",  # Sum of SPAR for all picks
+    "manager_avg_spar",  # Average SPAR per pick
+    "manager_hit_rate",  # % of picks with positive SPAR
+    "manager_picks_count",  # Total picks made
     # === NEW: Starter Designation ===
-    "position_draft_rank",          # Rank within position for this manager-year (QB1=1, QB2=2)
-    "position_draft_label",         # Label like "QB1", "RB2", "WR3"
-    "starter_slots_available",      # How many starter slots exist for this position
-    "drafted_as_starter",           # 1 if drafted to be a starter, 0 if backup
-    "drafted_as_backup",            # 1 if drafted to be a backup, 0 if starter
-
+    "position_draft_rank",  # Rank within position for this manager-year (QB1=1, QB2=2)
+    "position_draft_label",  # Label like "QB1", "RB2", "WR3"
+    "starter_slots_available",  # How many starter slots exist for this position
+    "drafted_as_starter",  # 1 if drafted to be a starter, 0 if backup
+    "drafted_as_backup",  # 1 if drafted to be a backup, 0 if starter
     # === NEW: Position Tiers ===
-    "position_tier",                # Dynamic tier number (varies by position-year)
-    "position_tier_label",          # Descriptive label ("Elite RB", "Starter WR", etc.)
-    "position_tier_count",          # How many tiers exist for this position-year
-    "position_percentile",          # Exact percentile rank (0-100) within position-year
-
+    "position_tier",  # Dynamic tier number (varies by position-year)
+    "position_tier_label",  # Descriptive label ("Elite RB", "Starter WR", etc.)
+    "position_tier_count",  # How many tiers exist for this position-year
+    "position_percentile",  # Exact percentile rank (0-100) within position-year
     # === NEW: Bench Insurance Metrics ===
-    "bench_insurance_discount",     # Position-specific bench value discount (0.0-1.0)
-    "bench_spar",                   # Pre-computed bench SPAR = max(0, manager_spar) * discount
-    "position_failure_rate",        # Starter failure rate for this position
-    "position_activation_rate",     # Bench activation rate for this position
-
+    "bench_insurance_discount",  # Position-specific bench value discount (0.0-1.0)
+    "bench_spar",  # Pre-computed bench SPAR = max(0, manager_spar) * discount
+    "position_failure_rate",  # Starter failure rate for this position
+    "position_activation_rate",  # Bench activation rate for this position
     # === NEW: Rank-Based Bench Values (data-driven) ===
-    "bench_value_by_rank",          # Median SPAR for this position draft rank (QB2, RB3, etc.)
-    "slot_median_spar",             # Historical median SPAR for this draft slot
+    "bench_value_by_rank",  # Median SPAR for this position draft rank (QB2, RB3, etc.)
+    "slot_median_spar",  # Historical median SPAR for this draft slot
 ]
 
 
@@ -142,8 +125,8 @@ def get_available_columns() -> list:
         schema_df = run_query(schema_query)
 
         # Get column names from schema
-        if 'column_name' in schema_df.columns:
-            available = set(schema_df['column_name'].tolist())
+        if "column_name" in schema_df.columns:
+            available = set(schema_df["column_name"].tolist())
         elif len(schema_df.columns) > 0:
             # First column usually has names
             available = set(schema_df.iloc[:, 0].tolist())
@@ -157,21 +140,47 @@ def get_available_columns() -> list:
     except Exception:
         # If schema query fails, return core columns that should always exist
         core_cols = [
-            "player", "yahoo_position", "position", "nfl_team", "manager", "year",
-            "cost", "pick", "round", "cost_bucket",
-            "is_keeper_status", "is_keeper_cost", "kept_next_year",
-            "total_fantasy_points", "season_ppg", "games_played",
-            "weeks_rostered", "weeks_started",
-            "player_spar", "manager_spar", "spar",
-            "player_ppg", "manager_ppg",
-            "player_pgvor", "manager_pgvor", "pgvor",
+            "player",
+            "yahoo_position",
+            "position",
+            "nfl_team",
+            "manager",
+            "year",
+            "cost",
+            "pick",
+            "round",
+            "cost_bucket",
+            "is_keeper_status",
+            "is_keeper_cost",
+            "kept_next_year",
+            "total_fantasy_points",
+            "season_ppg",
+            "games_played",
+            "weeks_rostered",
+            "weeks_started",
+            "player_spar",
+            "manager_spar",
+            "spar",
+            "player_ppg",
+            "manager_ppg",
+            "player_pgvor",
+            "manager_pgvor",
+            "pgvor",
             "replacement_ppg",
-            "draft_roi", "spar_per_dollar", "spar_per_pick", "spar_per_round",
-            "points_per_dollar", "points_per_pick",
-            "season_overall_rank", "season_position_rank",
-            "price_rank_within_position", "pick_rank_within_position",
-            "value_over_replacement", "draft_position_delta",
-            "pick_savings", "cost_savings",
+            "draft_roi",
+            "spar_per_dollar",
+            "spar_per_pick",
+            "spar_per_round",
+            "points_per_dollar",
+            "points_per_pick",
+            "season_overall_rank",
+            "season_position_rank",
+            "price_rank_within_position",
+            "pick_rank_within_position",
+            "value_over_replacement",
+            "draft_position_delta",
+            "pick_savings",
+            "cost_savings",
         ]
         return core_cols
 
@@ -223,8 +232,8 @@ def load_draft_data() -> Dict[str, Any]:
         df = run_query(query)
 
         # Rename total_fantasy_points to points for compatibility with existing code
-        if 'total_fantasy_points' in df.columns:
-            df = df.rename(columns={'total_fantasy_points': 'points'})
+        if "total_fantasy_points" in df.columns:
+            df = df.rename(columns={"total_fantasy_points": "points"})
 
         return {"Draft History": df}
 
@@ -258,14 +267,16 @@ def load_draft_optimizer_data(start_year: int, end_year: int) -> pd.DataFrame:
         """
         df = run_query(q)
         if df.empty:
-            return pd.DataFrame(columns=['primary_position', 'cost_bucket', 'cost', 'PPG'])
+            return pd.DataFrame(
+                columns=["primary_position", "cost_bucket", "cost", "PPG"]
+            )
 
         # Clean dtypes
-        df['cost'] = df['cost'].round(2).astype(float)
-        df['PPG'] = df['PPG'].round(2).astype(float)
-        df['cost_bucket'] = df['cost_bucket'].astype(int)
+        df["cost"] = df["cost"].round(2).astype(float)
+        df["PPG"] = df["PPG"].round(2).astype(float)
+        df["cost_bucket"] = df["cost_bucket"].astype(int)
 
         return df
     except Exception as e:
         st.error(f"Failed to load draft optimizer data: {e}")
-        return pd.DataFrame(columns=['primary_position', 'cost_bucket', 'cost', 'PPG'])
+        return pd.DataFrame(columns=["primary_position", "cost_bucket", "cost", "PPG"])
